+++ invisible_hand/scripts/grant_read_access.py
@@ -1,5 +1,4 @@
 import click
-from six import ensure_binary
 import trio
 from tqdm import tqdm
 
@@ -7,7 +6,7 @@ from typing import List
 
 from ..utils.github_scanner import query_matching_repos
 from ..utils.github_entities import Team
-from ..utils.github_api import ensure_gh_token
+from ..utils import ensure_gh_token
 from ..config.github import config_github, config_grant_read_access
 
 # Use a "team slug"
